backend/requirements.txt
alembic==1.17.2
amqp==5.3.1
asgiref==3.9.2
attrs==25.4.0
azure-core==1.37.0
azure-storage-blob==12.27.1
beautifulsoup4==4.14.3
billiard==4.2.2
blinker==1.9.0
boto3==1.42.12
botocore==1.42.12
celery==5.5.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2
django-auditlog==3.3.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.2
django-ipware==7.0.1
django-jazzmin==3.0.1
django-prometheus==2.4.1
django-storages==1.14.6
django-structlog==9.1.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
requests==2.32.5
s3transfer==0.16.0
selenium==4.39.0
sentry-sdk==2.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
structlog==25.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.4
whitenoise==6.11.0
wsproto==1.3.2

gunicorn==21.2.0

psycopg2-binary==2.9.9
